+++ misc/get_estimates_art_coverage_participants_vl200.R
@@ -4,14 +4,14 @@ library(ggplot2)
 library(scales)
 library(lubridate)
 library(rstan)
-library("haven")
+library(haven)
 library(here)
 
 # directory of the repository
 gitdir <- here()
 
 # load paths
-source(file.path(gitdir, "paths.R"))
+source(file.path(gitdir, "config.R"))
 
 # TODO: shozen: do you think this would be helpful? 
 # library(optparse)
@@ -153,7 +153,10 @@ for(round in c("R015", 'R016', 'R017', 'R018')){
   # run and save model
   fit <- sampling(stan.model, data=stan.data, iter=10e3, warmup=5e2, chains=1, control = list(max_treedepth= 15, adapt_delta= 0.999))
   filename <- paste0('art_gp_stanfit_round',gsub('R0', '', round),'_vl200_221208.rds')
-  saveRDS(fit, file=file.path(outdir,filename))
+  if( ! file.exists(filename) | config$overwrite.existing.files )
+  {
+    saveRDS(fit, file=file.path(outdir,filename))
+  }
   # fit <- readRDS(file.path(outdir,filename))
 }
 
@@ -315,7 +318,7 @@ stats[['max_rhat']] = convergence[, round(max(rhat), 4)]
 
 # file.name <- file.path(gitdir.fit,'RCCS_art_posterior_samples_vl200_221208.rds')
 file.name <- file.selfreportedart.vl200 
-if(! file.exists(file.name))
+if(! file.exists(file.name) | config$overwrite.existing.files )
 {
     cat("Saving file:", file.name, '\n')
     saveRDS(nsinf.samples, file = file.name)
@@ -324,4 +327,10 @@ if(! file.exists(file.name))
 }
 
 file.name <- file.path(outdir, 'RCCS_art_model_fit_vl200_221208.RDS')
-saveRDS(stats, file = file.name)
+if(! file.exists(file.name) | config$overwrite.existing.files )
+{
+    cat("Saving file:", file.name, '\n')
+    saveRDS(stats, file = file.name)
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
