+++ misc/get_estimates_unsuppressed_proportion_non_participants.R
@@ -6,7 +6,7 @@ library(here)
 
 # directory of the repository
 gitdir <- here()
-source(file.path(gitdir, "paths.R"))
+source(file.path(gitdir, "config.R"))
 
 # TODO: shozen: do you think this would be helpful? 
 # library(optparse)
@@ -309,7 +309,7 @@ stats[['max_rhat']] = convergence[, round(max(rhat), 4)]
 
 # file.name <- file.path(gitdir.fit, 'RCCS_nonsuppressed_proportion_posterior_samples_vl_1000_newlyregistered_221101.rds')
 file.name <- file.unsuppressedviralload.newly 
-if(! file.exists(file.name))
+if(! file.exists(file.name)  | config$overwrite.existing.files )
 {
     cat("Saving file:", file.name, '\n')
     saveRDS(nsinf.samples, file = file.name)
@@ -317,5 +317,13 @@ if(! file.exists(file.name))
     cat("File:", file.name, "already exists...\n")
 }
 
+
 file.name <- file.path(outdir, 'RCCS_nonsuppressed_proportion_model_fit_newlyregistered_221101.RDS')
-saveRDS(stats, file = file.name)
+if(! file.exists(file.name) | config$overwrite.existing.files )
+{
+    cat("Saving file:", file.name, '\n')
+    saveRDS(stats, file = file.name)
+
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
