+++ misc/get_estimates_unsuppressed_proportion_non_participants_vl200.R
@@ -6,7 +6,7 @@ library(here)
 
 # directory of the repository
 gitdir <- here()
-source(file.path(gitdir, "paths.R"))
+source(file.path(gitdir, "config.R"))
 
 # TODO: shozen: do you think this would be helpful? 
 # library(optparse)
@@ -78,7 +78,6 @@ if(1){
     scale_y_continuous(expand = expansion(mult = c(0, 0.05))) + 
     scale_x_continuous(expand = c(0,0))
   ggsave(p, file=file.path(outdir, paste0('count_unsuppressed_by_gender_loc_age_newlyregistered_vl200_221121.pdf')), w=7, h=5.2)
-  
 }
 
 
@@ -132,9 +131,11 @@ for(round in 15:18){
   # run and save model
   fit <- sampling(stan.model, data=stan.data, iter=10e3, warmup=5e2, chains=1, control = list(max_treedepth= 15, adapt_delta= 0.999))
   filename <- paste0( '220729f_notsuppAmongInfected_gp_stan_round',round,'_vl_200_newlyregistered.rds')
-  saveRDS(fit, file=file.path(outdir,filename))
-  # fit <- readRDS(file.path(outdir,filename))
-  
+  filename <- file.path(outdir, filename)
+  if( ! file.exists(filename) | config$overwrite.existing.files )
+  {
+      saveRDS(fit, filename)
+  }
 }
 
 
@@ -290,7 +291,7 @@ stats[['max_rhat']] = convergence[, round(max(rhat), 4)]
 
 # file.name <- file.path(gitdir.fit, 'RCCS_nonsuppressed_proportion_posterior_samples_vl_200_newlyregistered_221121.rds')
 file.name <- file.unsuppressedviralload.newly.vl200 
-if(! file.exists(file.name))
+if(! file.exists(file.name) | config$overwrite.existing.files)
 {
     cat("Saving file:", file.name, '\n')
     saveRDS(nsinf.samples, file = file.name)
@@ -299,4 +300,11 @@ if(! file.exists(file.name))
 }
 
 file.name <- file.path(outdir, 'RCCS_nonsuppressed_proportion_model_fit_newlyregistered_vl200_221121.RDS')
-saveRDS(stats, file = file.name)
+if(! file.exists(file.name) | config$overwrite.existing.files )
+{
+    cat("Saving file:", file.name, '\n')
+    saveRDS(stats, file = file.name)
+
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
